refactor: split openai-responses adapter into openai-responses/ folder
